restructured to abstract algorithm from pathfinder object
