+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pathfinder"
-version = "0.0.0"
+version = "1.0.0"
 authors = ["UAV Austin <uavaustin@gmail.com>"]
 license = "MIT"
